start on excel parsing. debugging
