+++ app/views/_navbar.erb
@@ -1,4 +1,4 @@
-<form action="/workouts" method="post"><input id='create-workout' type="submit" value="Log a workout"></form>
+<form action="/" method="post"><input id='create-workout' type="submit" value="Log a workout"></form>
 <form action="/workouts/show" method="get"><input id='all-workouts-nav' type="submit" value="View All Workouts"></form>
 <form action="/charts" method="get"><input id="go-to-charts" type="submit" value="Go to Charts"></form>
 <form action='/search' method="get"><input id="search-workouts" type="submit" value="Search Workouts"></form>
