+++ app/views/upload.erb
@@ -1,0 +1,10 @@
+<div class="container transbox">
+  <h2>Add Your Own Workout Data</h2>
+  <form action='workouts/upload' method = 'post'>
+    <label for="file">Upload your Excel file</label>
+    <input type="file" name="file">
+    <input type="submit" value="Upload">
+  </form>
+  <br>
+  <span class="waiting">Thank you for your patience, this may take a while</span>
+</div>
