+++ public/css/application.css
@@ -68,6 +68,12 @@ td {
 header {
   font-size: 3em;
   font-family: Verdana;
+  position: relative;
+  text-align: center;
+  box-shadow: 2px 2px 10px #333;
+  border-bottom-left-radius: 30px;
+  border-bottom-right-radius: 30px;
+  background-color: rgba(48, 144, 199, 0.2);
 }
 
 navbar > form {
@@ -109,6 +115,10 @@ input[type="submit"] {
 
 div.transbox {
   background-color: rgba(256, 256, 256, 0.6);
+}
+
+.waiting {
+  display: none;
 }
 /*debugger*/
 /** {
